requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
baidu-aip==4.16.13
beautifulsoup4==4.12.3
blobfile==2.1.1
bs4==0.0.2
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect==1.2.0
curl_cffi==0.6.2
deep-translator==1.11.4
distro==1.9.0
exceptiongroup==1.2.0
feedparser==6.0.11
filelock==3.13.4
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Session==0.8.0
frida==16.1.10
frida-tools==12.3.0
frozenlist==1.4.1
googletrans==3.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
libretranslatepy==2.1.1
loguru==0.6.0
lxml==4.9.4
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
nest-asyncio==1.6.0
openai==1.16.2
Pandora-ChatGPT @ file:///Users/s1xu/tempDir/pandora
Pandora-Cloud==0.7.3
pandora-tiktoken==0.3.1
prompt-toolkit==3.0.43
pyaes==1.6.1
pyasn1==0.5.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.6.0
pyperclip==1.8.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.3.5
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
tabulate==0.9.0
telegram==0.0.1
Telethon==1.34.0
tqdm==4.66.2
translate==3.6.1
translator==0.0.9
typing_extensions==4.9.0
tzlocal==5.2
urllib3==1.26.18
waitress==2.1.2
wcwidth==0.2.12
Werkzeug==2.2.3
yarl==1.9.4
yoyo-migrations==8.2.0
zipp==3.18.1
